Refactor allocateip file structure
